Important bug fix

+++ java-classes/package_downloader.java
@@ -28,10 +28,15 @@ public class package_downloader extends MaxObject {
     File to = new File(outputfolder + File.separator + packagename);
     
     //if package was already there, delete it
-    if (to.exists()){
-      deleteFolder(to);
+    if (!rootFolder.substring(0,rootFolder.length()-1).equals(packagename))
+    {
+        if (to.exists())
+        {
+          deleteFolder(to);
+        }
+        from.renameTo(to);
     }
-    from.renameTo(to);
+
 
     //if package was in subfolder, delete parent folder
     if (!relativepath.equals("None"))
@@ -124,6 +129,10 @@ public class package_downloader extends MaxObject {
    return rootFolder;  
    }  
 }
+
+
+
+
